Create max_area_of_island.cpp
You are given an m x n binary matrix grid. An island is a group of 1's (representing land) connected 4-directionally (horizontal or vertical.) You may assume all four edges of the grid are surrounded by water.

The area of an island is the number of cells with a value of 1 on the island.

Return the maximum area of an island in the grid. If there is no island, return 0.